requirements.txt
defusedxml==0.4.1
dj-database-url==0.4.1
Django==1.9.7
django-appconf==1.0.2
django-compressor==2.0
oauthlib==1.1.2
PyJWT==1.4.0
python-social-auth==0.2.19
python3-openid==3.0.10
rcssmin==1.0.6
requests==2.10.0
requests-oauthlib==0.6.1
rjsmin==1.0.12
six==1.10.0

gunicorn==19.6.0
psycopg2==2.6.1
whitenoise==3.2
